+++ src/Input/Path.hpp
@@ -30,35 +30,36 @@ public:
     
 
     // return the left bound of the interval defined by this Path (in samples)
-    int begin() const;
+    unsigned int begin() const;
 
     // return the right bound of the interval defined by this Path (in samples)
     // WARNING: the right bound is the last position of the interval +1
     // i.e. it does not belong to the interval
-    int end() const;
+    unsigned int end() const;
 
     // return the length of the interval defined by this Path (in samples)
-    int length() const;
+    unsigned int length() const;
 
     // the position in param belongs to the interval defined by this Path
-    bool member(int);
+    bool member(unsigned int);
 
     // the position in param belongs to the interval defined by this Path
     // and is closer to the left bound than to the right bound
-    bool aligned(int);
+    bool aligned(unsigned int);
     
     // sub(n, i) return a new path defined as the ith subsegment
     // of a division of this Path into n segments
     // the interval length must be divisible by n
-    Path* sub(int, int);
+    // UNUSED? TODO REM.
+    Path* sub(unsigned int, unsigned int);
     
 protected:
     
     // position of left bound of interval (in samples)
-    int _begin;
+    unsigned int _begin;
     
     // length of interval (in samples)
-    int _len;
+    unsigned int _len;
     
     // _dur : inverse of duration defined by path
 };
@@ -81,7 +82,7 @@ public:
     // to the input segment s.
     // construction of recursive paths.
     // WARNING: must be aligned.
-    Alignment(Segment*, int, int);
+    Alignment(Segment*, unsigned int, unsigned int);
     
     ~Alignment();
     
@@ -92,49 +93,50 @@ public:
     //int date(size_t) const;
     
     // number of points of segment in the fist half of this interval
-    int l_size() const { return _seg_llen; }
+    unsigned int l_size() const { return _seg_llen; }
     
     // index in segment of the first element of segment inside the fist half of this interval
     // out_of_range (= size of segment) id lsize() == 0
-    size_t l_first() const { return _seg_lbeg; }
+    unsigned int l_first() const { return _seg_lbeg; }
 
     // number of points of segment in the second half of this interval
-    int r_size() const { return _seg_rlen; }
+    unsigned int r_size() const { return _seg_rlen; }
     
     // index in segment of the first element of segment inside the second half of this interval
     // out_of_range (= size of segment) id rsize() == 0
-    size_t r_first() const { return _seg_rbeg; }
+    unsigned int r_first() const { return _seg_rbeg; }
     
     // sub(k) return the list of subpaths obtained by
     // division of this Path into k segments.
+    // k must be > 1.
     // the interval length must be divisible by k.
-    vector<Alignment*> subs(int);
+    vector<Alignment*> subs(unsigned int);
     
 private:
 
     // resolution
-    int _res;  // SUPPR (deja dans segment)
+    unsigned int _res;  // SUPPR (deja dans segment)
     
     Segment* _seg;
     
     // number of segment points in the first half of this path
-    int _seg_llen;
+    unsigned int _seg_llen;
 
     // index of first segment point in the first half of this path
     // or out_of_bound (segment size) is there are none
-    size_t _seg_lbeg;
+    unsigned int _seg_lbeg;
     
     // number of segment points in the second half of this path
-    int _seg_rlen;
+    unsigned int _seg_rlen;
 
     // index of first segment point in the second half of this path
     // or out_of_bound (segment size) is there are none
-    size_t _seg_rbeg;
+    unsigned int _seg_rbeg;
     
     
     // align(b) set the above values, starting from index b
     // and return the next index in segment to be processed
     // or the size of segment if end of segment is reached.
-    size_t align(size_t);
+    unsigned int align(unsigned int);
     
 };
